% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Check for Absence}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{Object to check}

\item{y}{Object to check}
}
\value{
Invisible
}
\description{
\code{\%ni\%}          Not in
}
\examples{
10 \%ni\% 1:9

}
