% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{det_pck}
\alias{det_pck}
\title{Package detacher}
\usage{
det_pck(f_pkg, f_char.only = FALSE)
}
\arguments{
\item{f_pkg}{Package name}

\item{f_char.only}{Package name handler}
}
\value{
Invisible
}
\description{
Detaches specified packages
}
\examples{
det_pck("data.table")
}
