% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_pck.R
\name{detPck}
\alias{detPck}
\title{Package detacher}
\usage{
detPck(fPkg, fChar.only = FALSE)
}
\arguments{
\item{fPkg}{Package name}

\item{fChar.only}{Package name handler}
}
\value{
Invisible
}
\description{
Detaches specified packages
}
\examples{
detPck("data.table")
}
