% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{show}
\alias{show}
\alias{wideScreen}
\title{Change Display options}
\usage{
show(tbl_df)

wideScreen(howWide = NULL)
}
\arguments{
\item{tbl_df}{Tibble to display}

\item{howWide}{Number of columns to widen the screen. Default system obtained measures for "full screen".}
}
\value{
Invisible

Invisible
}
\description{
\code{show}          Show all tibble rows
}
\examples{
show(tibble::tibble(iris))

wideScreen()

}
