% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{doc}
\alias{doc}
\title{Choose documentation viewer}
\usage{
doc(str = NULL)
}
\arguments{
\item{str}{Viewer}
}
\value{
Invisible
}
\description{
\code{doc}           Documentation viewer
}
\examples{
doc('Opera')

}
