% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.R
\name{installed}
\alias{installed}
\alias{isInstalled}
\title{Check for Installed packages}
\usage{
installed()

isInstalled(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
List of packages

Boolean
}
\description{
\code{installed} - List installed packages
}
\examples{
installed()

isInstalled('stats')

}
