% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInstalled.R
\name{isInstalled}
\alias{isInstalled}
\title{Installed package finder}
\usage{
isInstalled(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
Boolean
}
\description{
Query whether package is installed
}
\examples{
       isInstalled('SystWrapper')
}
