% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{lsm}
\alias{lsm}
\title{Methods lister}
\usage{
lsm(f_object)
}
\arguments{
\item{f_object}{Object to inquiry about}
}
\value{
Object's methods
}
\description{
List object methods
}
\examples{
x <- 1:100; lsm(x)
}
