% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsMethods.R
\name{lsMethods}
\alias{lsMethods}
\title{Methods lister}
\usage{
lsMethods(object)
}
\arguments{
\item{object}{Object to inquiry about}
}
\value{
Object's methods
}
\description{
List object methods
}
\examples{
lsMethods(1:1000)
}
