% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp.R
\name{lsp}
\alias{lsp}
\title{Package lister}
\usage{
lsp(fPackage, fAll.names = FALSE, fPattern)
}
\arguments{
\item{fPackage}{Package name}

\item{fAll.names}{Boolean. Default = F}

\item{fPattern}{Specify if a pattern is to be searched}
}
\value{
Package contents
}
\description{
List package contents
}
\examples{
lsp(stats)
}
