% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsPkg.R
\name{lsPkg}
\alias{lsPkg}
\title{Package lister}
\usage{
lsPkg(package, all_names = FALSE, pattern)
}
\arguments{
\item{package}{Package name}

\item{all_names}{Boolean. Default = F}

\item{pattern}{Specify if a pattern is to be searched}
}
\value{
Package contents
}
\description{
List package contents
}
\examples{
       lsPkg(stats)
}
