% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_swb.R
\name{normal_swb}
\alias{normal_swb}
\title{Normal Soil Water Balance}
\usage{
normal_swb(month, ET0, P, data, awc, start_month)
}
\arguments{
\item{month}{A string that identifies the column of the month}

\item{ET0}{A string that identifies the column of the reference evapotranspiration}

\item{P}{A string that identifies the column of the precipitation}

\item{data}{A data.frame with (at least) three columns: month, reference evapotranspiration (et0) and precipitation (p)}

\item{awc}{Available water capacity of the soil}

\item{start_month}{Month of beginning of the soil water balance}
}
\description{
Calculates the Normal (climatological) soil water balance for a 12 month period.
}
\examples{
\dontrun{
swb_franca_sp <- tibble::tribble(
  ~m, ~et, ~p,
  "Jan", 117, 275,
  "Feb", 102, 218,
  "Mar", 104, 180,
  "Apr", 79, 60,
  "May", 60, 25,
  "Jun", 49, 20,
  "Jul", 54, 15,
  "Aug", 74, 12,
  "Sep", 93, 48,
  "Oct", 107, 113,
  "Nov", 108, 180,
  "Dec", 117, 245
)
normal_swb(month = "m", ET0 = "et", P = "p", data = swb_franca_sp, awc = 125, start_month = "Apr")
}

}
