% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smith_waterman.r
\name{smithwaterman}
\alias{smithwaterman}
\title{Smith - Waterman}
\usage{
smithwaterman(first, second, match = 3, missmatch = -3, gap = -1)
}
\arguments{
\item{first}{is the first sequence in the smith waterman algorithm}

\item{second}{is the second sequence in the smith waterman algorithm}

\item{match}{is the match score in the smith waterman algorithm}

\item{missmatch}{is the missmatch score in the smith waterman algorithm}

\item{gap}{is the gap score in the smith waterman algorithm}
}
\value{
The optimal alignment between the first sequence and the second,
        a result variable that contains a rappresentation of the differences
}
\description{
Returns one of the optimal local aligments between two sequences
}
\examples{
optimal_alignment <- smithwaterman(DNAString("ACCTG"), DNAString("ATCTG"), 3, -3, -1);
optimal_alignment <- smithwaterman(DNAString("ACCTG"), DNAString("ATCTG"), 3, -1, 0);
}
