% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideScreen.R
\name{wideScreen}
\alias{wideScreen}
\title{Screen wider}
\usage{
wideScreen(howWide = NULL)
}
\arguments{
\item{howWide}{Number of columns to widen the screen. Default system obtained measures for "full screen".}
}
\value{
Invisible
}
\description{
Use full screen on interactive R sessions
}
\examples{
       wideScreen()
}
