% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaded.R
\name{loaded}
\alias{loaded}
\alias{isloaded}
\title{Check for Loaded Packages}
\usage{
loaded(..., all = FALSE, char, character.only = FALSE)

isloaded(...)
}
\arguments{
\item{all}{logical.  If \code{TRUE} will show all packages
including base install; \code{FALSE} will show all packages
excluding base install packages that install when R loads.}

\item{char}{Character vector containing packages to load.  If you are calling
\code{loaded} from within a function (or just having difficulties calling it
using a character vector input) then pass your character vector of packages
to load to this parameter directly.}

\item{character.only}{logical. If \code{TRUE} then \code{loaded} will only
accept a single input which is a character vector containing the names of
packages to load.}

\item{\ldots}{Optional package names.  Adding package names will check their
individual load status.}
}
\description{
\code{loaded} - Output is a character string of loaded packages.

\code{isloaded} - Check if package(s) is loaded.
}
