% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp.R
\name{lsp}
\alias{lsp}
\title{Package lister}
\usage{
lsp(f_package, f_all.names = FALSE, f_pattern)
}
\arguments{
\item{f_package}{Package name}

\item{f_all.names}{Boolean. Default = F}

\item{f_pattern}{Specify if a pattern is to be searched}
}
\value{
Package contents
}
\description{
List package contents
}
\examples{
lsp(stats)
}
